fix: separate styles from components
