+++ app/(tabs)/index.tsx
@@ -1,56 +1,39 @@
 import React, { useState } from 'react';
-import { FlatList, StyleSheet, Pressable, Modal } from 'react-native';
+import { FlatList, Pressable } from 'react-native';
 import { useAppContext } from '@/hooks/AppContext';
 import { ThemedText } from '@/components/common/ThemedText';
 import { ThemedView } from '@/components/common/ThemedView';
 import { ChatListItem } from '@/components/chat/ChatListItem';
-import { UserListItem } from '@/components/UserListItem';
 import { IconSymbol } from '@/components/ui/IconSymbol';
 import { SearchModal } from '@/components/chat/SearchModal';
+import { NewChatModal } from '@/components/chat/NewChatModal';
+import { chatsScreenStyles } from '@/styles/screens/chatsScreen.styles';
 
 export default function ChatsScreen() {
-  const { currentUser, users, chats, createChat, theme } = useAppContext();
+  const { currentUser, chats, createChat, theme } = useAppContext();
   const [modalVisible, setModalVisible] = useState(false);
   const [searchVisible, setSearchVisible] = useState(false);
-  const [selectedUsers, setSelectedUsers] = useState<string[]>([]);
-
-  const toggleUserSelection = (userId: string) => {
-    if (selectedUsers.includes(userId)) {
-      setSelectedUsers(selectedUsers.filter(id => id !== userId));
-    } else {
-      setSelectedUsers([...selectedUsers, userId]);
-    }
-  };
-
-  const handleCreateChat = () => {
-    if (currentUser && selectedUsers.length > 0) {
-      const participants = [currentUser.id, ...selectedUsers];
-      createChat(participants);
-      setModalVisible(false);
-      setSelectedUsers([]);
-    }
-  };
 
   const renderEmptyComponent = () => (
-    <ThemedView style={styles.emptyContainer}>
-      <ThemedText style={styles.emptyText}>No chats yet</ThemedText>
+    <ThemedView style={chatsScreenStyles.emptyContainer}>
+      <ThemedText style={chatsScreenStyles.emptyText}>No chats yet</ThemedText>
       <ThemedText>Tap the + button to start a new conversation</ThemedText>
     </ThemedView>
   );
 
   return (
-    <ThemedView style={styles.container}>
-      <ThemedView style={styles.header}>
+    <ThemedView style={chatsScreenStyles.container}>
+      <ThemedView style={chatsScreenStyles.header}>
         <ThemedText type="title">Chats</ThemedText>
-        <ThemedView style={styles.headerButtons}>
+        <ThemedView style={chatsScreenStyles.headerButtons}>
           <Pressable
-            style={styles.headerButton}
+            style={chatsScreenStyles.headerButton}
             onPress={() => setSearchVisible(true)}
           >
             <IconSymbol name="search" size={24} color={theme.colors.tint} />
           </Pressable>
           <Pressable
-            style={styles.headerButton}
+            style={chatsScreenStyles.headerButton}
             onPress={() => setModalVisible(true)}
           >
             <IconSymbol name="add" size={24} color={theme.colors.tint} />
@@ -68,63 +51,14 @@ export default function ChatsScreen() {
           />
         )}
         ListEmptyComponent={renderEmptyComponent}
-        contentContainerStyle={styles.listContainer}
+        contentContainerStyle={chatsScreenStyles.listContainer}
       />
 
-      <Modal
-        animationType="fade"
-        transparent={true}
+      <NewChatModal
         visible={modalVisible}
-        onRequestClose={() => {
-          setModalVisible(false);
-          setSelectedUsers([]);
-        }}
-      >
-        <ThemedView style={styles.modalContainer}>
-          <ThemedView style={[styles.modalContent, { backgroundColor: theme.colors.backgroundChat }]}>
-            <ThemedView style={[styles.modalHeader, { backgroundColor: theme.colors.backgroundChat }]}>
-              <ThemedText type="subtitle">New Chat</ThemedText>
-              <Pressable onPress={() => {
-                setModalVisible(false);
-                setSelectedUsers([]);
-              }}>
-                <IconSymbol name="close" size={24} color={theme.colors.error} />
-              </Pressable>
-            </ThemedView>
-
-            <ThemedText style={styles.modalSubtitle}>
-              Select users to chat with
-            </ThemedText>
-
-            <FlatList
-              data={users.filter(user => user.id !== currentUser?.id)}
-              keyExtractor={(item) => item.id}
-              renderItem={({ item }) => (
-                <UserListItem
-                  user={item}
-                  onSelect={() => toggleUserSelection(item.id)}
-                  isSelected={selectedUsers.includes(item.id)}
-                />
-              )}
-              style={styles.userList}
-            />
-
-            <Pressable
-              style={[
-                styles.createButton,
-                { backgroundColor: theme.colors.tint },
-                selectedUsers.length === 0 && styles.disabledButton
-              ]}
-              onPress={handleCreateChat}
-              disabled={selectedUsers.length === 0}
-            >
-              <ThemedText style={[styles.createButtonText, { color: theme.colors.background, }]}>
-                Create Chat
-              </ThemedText>
-            </Pressable>
-          </ThemedView>
-        </ThemedView>
-      </Modal>
+        onClose={() => setModalVisible(false)}
+        onCreateChat={createChat}
+      />
 
       <SearchModal
         visible={searchVisible}
@@ -133,86 +67,3 @@ export default function ChatsScreen() {
     </ThemedView>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    paddingTop: 60,
-  },
-  header: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    paddingHorizontal: 20,
-    paddingBottom: 20,
-  },
-  headerButtons: {
-    flexDirection: 'row',
-    gap: 16,
-  },
-  headerButton: {
-    width: 40,
-    height: 40,
-    borderRadius: 20,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: 'rgba(0, 122, 255, 0.1)',
-  },
-  listContainer: {
-    flexGrow: 1,
-  },
-  emptyContainer: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    padding: 20,
-    marginTop: 40,
-  },
-  emptyText: {
-    fontSize: 18,
-    fontWeight: 'bold',
-    marginBottom: 10,
-  },
-  modalContainer: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: 'rgba(0, 0, 0, 0.5)',
-  },
-  modalContent: {
-    width: '90%',
-    maxHeight: '80%',
-    borderRadius: 10,
-    padding: 20,
-    elevation: 5,
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.25,
-    shadowRadius: 3.84,
-  },
-  modalHeader: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    marginBottom: 20,
-  },
-  modalSubtitle: {
-    marginBottom: 10,
-  },
-  userList: {
-    maxHeight: 400,
-  },
-  createButton: {
-    padding: 15,
-    borderRadius: 10,
-    alignItems: 'center',
-    marginTop: 20,
-  },
-  disabledButton: {
-    backgroundColor: '#CCCCCC',
-  },
-  createButtonText: {
-    // color: 'white',
-    fontWeight: 'bold',
-  },
-});
