+++ app/(tabs)/profile.tsx
@@ -1,10 +1,11 @@
 import React from 'react';
-import { StyleSheet, Pressable, SafeAreaView } from 'react-native';
+import { Pressable, SafeAreaView } from 'react-native';
 import { useAppContext } from '@/hooks/AppContext';
 import { ThemedText } from '@/components/common/ThemedText';
 import { ThemedView } from '@/components/common/ThemedView';
 import { Avatar } from '@/components/Avatar';
 import { IconSymbol } from '@/components/ui/IconSymbol';
+import { profileScreenStyles } from '@/styles/screens/profileScreen.styles';
 
 export default function ProfileScreen() {
   const { currentUser, logout } = useAppContext();
@@ -16,41 +17,41 @@ export default function ProfileScreen() {
 
   if (!currentUser) {
     return (
-      <ThemedView style={styles.loadingContainer}>
+      <ThemedView style={profileScreenStyles.loadingContainer}>
         <ThemedText>Loading user profile...</ThemedText>
       </ThemedView>
     );
   }
 
   return (
-    <SafeAreaView style={styles.safeArea}>
-      <ThemedView style={styles.container}>
-        <ThemedView style={styles.profileHeader}>
+    <SafeAreaView style={profileScreenStyles.safeArea}>
+      <ThemedView style={profileScreenStyles.container}>
+        <ThemedView style={profileScreenStyles.profileHeader}>
           <Avatar uri={currentUser.avatar} fallback={currentUser.name[0]} size={100} />
-          <ThemedView style={styles.profileInfo}>
+          <ThemedView style={profileScreenStyles.profileInfo}>
             <ThemedText type="title">{currentUser.name}</ThemedText>
-            <ThemedText style={styles.statusText}>
+            <ThemedText style={profileScreenStyles.statusText}>
               {currentUser.status.charAt(0).toUpperCase() + currentUser.status.slice(1)}
             </ThemedText>
           </ThemedView>
         </ThemedView>
-        
-        <ThemedView style={styles.section}>
+
+        <ThemedView style={profileScreenStyles.section}>
           <ThemedText type="subtitle">Account Information</ThemedText>
-          <ThemedView style={styles.infoRow}>
-            <ThemedText style={styles.infoLabel}>ID:</ThemedText>
+          <ThemedView style={profileScreenStyles.infoRow}>
+            <ThemedText style={profileScreenStyles.infoLabel}>ID:</ThemedText>
             <ThemedText>{currentUser.id}</ThemedText>
           </ThemedView>
 
-          <ThemedView style={styles.infoRow}>
-            <ThemedText style={styles.infoLabel}>Full Name:</ThemedText>
+          <ThemedView style={profileScreenStyles.infoRow}>
+            <ThemedText style={profileScreenStyles.infoLabel}>Full Name:</ThemedText>
             <ThemedText>{currentUser.name}</ThemedText>
           </ThemedView>
         </ThemedView>
-        <ThemedView style={styles.buttonContainer}>
-          <Pressable style={styles.logoutButton} onPress={handleLogout}>
+        <ThemedView style={profileScreenStyles.buttonContainer}>
+          <Pressable style={profileScreenStyles.logoutButton} onPress={handleLogout}>
             <IconSymbol name="log-out" size={20} color="#FFFFFF" />
-            <ThemedText style={styles.logoutText}>Log Out</ThemedText>
+            <ThemedText style={profileScreenStyles.logoutText}>Log Out</ThemedText>
           </Pressable>
         </ThemedView>
       </ThemedView>
@@ -58,63 +59,3 @@ export default function ProfileScreen() {
   );
 }
 
-const styles = StyleSheet.create({
-  safeArea: {
-    flex: 1,
-  },
-  container: {
-    flex: 1,
-    paddingTop: 60,
-  },
-  loadingContainer: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-    padding: 20,
-  },
-  profileHeader: {
-    alignItems: 'center',
-    padding: 20,
-  },
-  profileInfo: {
-    alignItems: 'center',
-    marginTop: 16,
-  },
-  statusText: {
-    fontSize: 16,
-    color: '#8F8F8F',
-    marginTop: 4,
-  },
-  section: {
-    padding: 20,
-    marginTop: 20,
-  },
-  infoRow: {
-    flexDirection: 'row',
-    marginTop: 12,
-  },
-  infoLabel: {
-    fontWeight: 'bold',
-    marginRight: 10,
-    width: 100,
-  },
-  buttonContainer: {
-    flex: 1,
-    justifyContent: 'flex-end',
-    marginBottom: 80, // Add padding to ensure the button is visible above the tab bar
-  },
-  logoutButton: {
-    flexDirection: 'row',
-    backgroundColor: '#FF3B30',
-    padding: 15,
-    borderRadius: 10,
-    alignItems: 'center',
-    justifyContent: 'center',
-    marginHorizontal: 20,
-  },
-  logoutText: {
-    color: 'white',
-    fontWeight: 'bold',
-    marginLeft: 10,
-  },
-});
