+++ app/login.tsx
@@ -1,11 +1,12 @@
 import React from 'react';
-import { StyleSheet, FlatList, SafeAreaView } from 'react-native';
+import { FlatList, SafeAreaView } from 'react-native';
 import { useRouter } from 'expo-router';
 import { StatusBar } from 'expo-status-bar';
 import { useAppContext } from '@/hooks/AppContext';
 import { ThemedText } from '@/components/common/ThemedText';
 import { ThemedView } from '@/components/common/ThemedView';
-import { UserListItem } from '@/components/UserListItem';
+import { UserListItem } from '@/components/chat/UserListItem';
+import { loginScreenStyles } from '@/styles/screens/loginScreen.styles';
 
 export default function LoginScreen() {
   const { users, login } = useAppContext();
@@ -18,16 +19,16 @@ export default function LoginScreen() {
   };
 
   return (
-    <SafeAreaView style={styles.safeArea}>
+    <SafeAreaView style={loginScreenStyles.safeArea}>
       <StatusBar style="auto" />
-      <ThemedView style={styles.container}>
-        <ThemedView style={styles.header}>
+      <ThemedView style={loginScreenStyles.container}>
+        <ThemedView style={loginScreenStyles.header}>
           <ThemedText type="title">Welcome to Chat App</ThemedText>
-          <ThemedText style={styles.subtitle}>
+          <ThemedText style={loginScreenStyles.subtitle}>
             Select a user to continue
           </ThemedText>
         </ThemedView>
-        
+
         <FlatList
           data={users}
           keyExtractor={(item) => item.id}
@@ -37,32 +38,10 @@ export default function LoginScreen() {
               onSelect={() => handleUserSelect(item.id)}
             />
           )}
-          contentContainerStyle={styles.listContainer}
+          contentContainerStyle={loginScreenStyles.listContainer}
         />
       </ThemedView>
     </SafeAreaView>
   );
 }
 
-const styles = StyleSheet.create({
-  safeArea: {
-    flex: 1,
-  },
-  container: {
-    flex: 1,
-    paddingTop: 20,
-  },
-  header: {
-    alignItems: 'center',
-    padding: 20,
-    marginBottom: 20,
-  },
-  subtitle: {
-    marginTop: 10,
-    fontSize: 16,
-    color: '#8F8F8F',
-  },
-  listContainer: {
-    paddingBottom: 20,
-  },
-});
