+++ components/chat/MediaPickerModal.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
-import { View, StyleSheet, Pressable, ActionSheetIOS, Platform, Modal } from 'react-native';
+import { View, Pressable, ActionSheetIOS, Platform, Modal } from 'react-native';
 import { ThemedText } from '@/components/common/ThemedText';
 import { IconSymbol } from '@/components/ui/IconSymbol';
 import { useColorScheme } from '@/hooks/useColorScheme';
-
+import { mediaPickerModalStyles } from '@/styles/components/mediaPickerModal.styles';
+import { MEDIA_OPTIONS } from '@/constants/mediaOptions';
 interface MediaPickerModalProps {
   visible: boolean;
   onClose: () => void;
@@ -11,12 +12,6 @@ interface MediaPickerModalProps {
   onSelectVideo: () => void;
   onSelectFile: () => void;
 }
-
-const MEDIA_OPTIONS = [
-  { icon: 'image-outline', label: 'Imagen', action: 'onSelectImage' },
-  { icon: 'videocam-outline', label: 'Video', action: 'onSelectVideo' },
-  { icon: 'document-outline', label: 'Archivo', action: 'onSelectFile' },
-] as const;
 
 export function MediaPickerModal({
   visible,
@@ -80,19 +75,19 @@ export function MediaPickerModal({
       animationType="fade"
       onRequestClose={onClose}
     >
-      <Pressable style={styles.overlay} onPress={onClose}>
+      <Pressable style={mediaPickerModalStyles.overlay} onPress={onClose}>
         <View
-          style={[styles.container, { backgroundColor: isDark ? '#2A2C33' : '#FFFFFF' }]}
+          style={[mediaPickerModalStyles.container, { backgroundColor: isDark ? '#2A2C33' : '#FFFFFF' }]}
           onStartShouldSetResponder={() => true}
         >
-          <ThemedText style={styles.title}>Select media</ThemedText>
-          <View style={styles.optionsContainer}>
+          <ThemedText style={mediaPickerModalStyles.title}>Select media</ThemedText>
+          <View style={mediaPickerModalStyles.optionsContainer}>
             {MEDIA_OPTIONS.map((option, index) => (
               <Pressable
                 key={index}
                 style={({ pressed }) => [
-                  styles.option,
-                  pressed && styles.optionPressed
+                  mediaPickerModalStyles.option,
+                  pressed && mediaPickerModalStyles.optionPressed
                 ]}
                 onPress={() => handleOptionPress(option.action)}
                 hitSlop={10}
@@ -102,7 +97,7 @@ export function MediaPickerModal({
                   size={24}
                   color={isDark ? '#FFFFFF' : '#000000'}
                 />
-                <ThemedText style={styles.optionText}>
+                <ThemedText style={mediaPickerModalStyles.optionText}>
                   {option.label}
                 </ThemedText>
               </Pressable>
@@ -114,44 +109,3 @@ export function MediaPickerModal({
   );
 }
 
-const styles = StyleSheet.create({
-  overlay: {
-    flex: 1,
-    backgroundColor: 'rgba(0, 0, 0, 0.5)',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  container: {
-    width: '80%',
-    borderRadius: 16,
-    padding: 20,
-    elevation: 5,
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.25,
-    shadowRadius: 4,
-  },
-  title: {
-    fontSize: 18,
-    fontWeight: 'bold',
-    marginBottom: 20,
-    textAlign: 'center',
-  },
-  optionsContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-around',
-  },
-  option: {
-    alignItems: 'center',
-    padding: 12,
-    borderRadius: 12,
-    minWidth: 80,
-  },
-  optionPressed: {
-    opacity: 0.7,
-  },
-  optionText: {
-    marginTop: 8,
-    fontSize: 14,
-  },
-});
