+++ components/chat/MessageStatus.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
-import { View, StyleSheet } from 'react-native';
+import { View} from 'react-native';
 import { IconSymbol } from '@/components/ui/IconSymbol';
 import { Message } from '@/interfaces/chatTypes';
 import { useColorScheme } from '@/hooks/useColorScheme';
+import { messageStatusStyles } from '@/styles/components/messageStatus.styles';
 
 interface MessageStatusProps {
   message: Message;
@@ -13,7 +14,7 @@ export function MessageStatus({ message }: MessageStatusProps) {
   const color = message.status === 'read' ? '#0a7ea4' : (isDark ? '#8E8E93' : '#8F8F8F');
 
   return (
-    <View style={styles.container}>
+    <View style={messageStatusStyles.container}>
       {message.status === 'sent' ? (
         <IconSymbol
           name="checkmark"
@@ -21,18 +22,18 @@ export function MessageStatus({ message }: MessageStatusProps) {
           color={color}
         />
       ) : (
-        <View style={styles.doubleCheckContainer}>
+        <View style={messageStatusStyles.doubleCheckContainer}>
           <IconSymbol
             name="checkmark"
             size={14}
             color={color}
-            style={styles.firstCheck}
+            style={messageStatusStyles.firstCheck}
           />
           <IconSymbol
             name="checkmark"
             size={14}
             color={color}
-            style={styles.secondCheck}
+            style={messageStatusStyles.secondCheck}
           />
         </View>
       )}
@@ -40,24 +41,3 @@ export function MessageStatus({ message }: MessageStatusProps) {
   );
 }
 
-const styles = StyleSheet.create({
-  container: {
-    marginLeft: 4,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  doubleCheckContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    width: 16,
-    height: 14,
-  },
-  firstCheck: {
-    position: 'absolute',
-    left: 0,
-  },
-  secondCheck: {
-    position: 'absolute',
-    left: 4,
-  },
-});
