+++ components/common/ThemedText.tsx
@@ -1,5 +1,5 @@
-import { Text, type TextProps, StyleSheet } from 'react-native';
-
+import { Text, type TextProps } from 'react-native';
+import { themedTextStyles } from '@/styles/common/themedText.styles';
 import { useThemeColor } from '@/hooks/useThemeColor';
 
 export type ThemedTextProps = TextProps & {
@@ -21,11 +21,11 @@ export function ThemedText({
     <Text
       style={[
         { color },
-        type === 'default' ? styles.default : undefined,
-        type === 'title' ? styles.title : undefined,
-        type === 'defaultSemiBold' ? styles.defaultSemiBold : undefined,
-        type === 'subtitle' ? styles.subtitle : undefined,
-        type === 'link' ? styles.link : undefined,
+        type === 'default' ? themedTextStyles.default : undefined,
+        type === 'title' ? themedTextStyles.title : undefined,
+        type === 'defaultSemiBold' ? themedTextStyles.defaultSemiBold : undefined,
+        type === 'subtitle' ? themedTextStyles.subtitle : undefined,
+        type === 'link' ? themedTextStyles.link : undefined,
         style,
       ]}
       {...rest}
@@ -33,28 +33,4 @@ export function ThemedText({
   );
 }
 
-const styles = StyleSheet.create({
-  default: {
-    fontSize: 16,
-    lineHeight: 24,
-  },
-  defaultSemiBold: {
-    fontSize: 16,
-    lineHeight: 24,
-    fontWeight: '600',
-  },
-  title: {
-    fontSize: 32,
-    fontWeight: 'bold',
-    lineHeight: 32,
-  },
-  subtitle: {
-    fontSize: 20,
-    fontWeight: 'bold',
-  },
-  link: {
-    lineHeight: 30,
-    fontSize: 16,
-    color: '#0a7ea4',
-  },
-});
+
