+++ constants/mediaOptions.ts
@@ -1,0 +1,5 @@
+export const MEDIA_OPTIONS = [
+    { icon: 'image-outline', label: 'Imagen', action: 'onSelectImage' },
+    { icon: 'videocam-outline', label: 'Video', action: 'onSelectVideo' },
+    { icon: 'document-outline', label: 'Archivo', action: 'onSelectFile' },
+] as const;
