+++ styles/common/themedText.styles.ts
@@ -1,0 +1,27 @@
+import { StyleSheet } from 'react-native';
+
+export const themedTextStyles = StyleSheet.create({
+    default: {
+      fontSize: 16,
+      lineHeight: 24,
+    },
+    defaultSemiBold: {
+      fontSize: 16,
+      lineHeight: 24,
+      fontWeight: '600',
+    },
+    title: {
+      fontSize: 32,
+      fontWeight: 'bold',
+      lineHeight: 32,
+    },
+    subtitle: {
+      fontSize: 20,
+      fontWeight: 'bold',
+    },
+    link: {
+      lineHeight: 30,
+      fontSize: 16,
+      color: '#0a7ea4',
+    },
+});
