+++ styles/components/mediaPickerModal.styles.ts
@@ -1,0 +1,43 @@
+import { StyleSheet } from 'react-native';
+
+export const mediaPickerModalStyles = StyleSheet.create({
+  overlay: {
+    flex: 1,
+    backgroundColor: 'rgba(0, 0, 0, 0.5)',
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  container: {
+    width: '80%',
+    borderRadius: 16,
+    padding: 20,
+    elevation: 5,
+    shadowColor: '#000',
+    shadowOffset: { width: 0, height: 2 },
+    shadowOpacity: 0.25,
+    shadowRadius: 4,
+  },
+  title: {
+    fontSize: 18,
+    fontWeight: 'bold',
+    marginBottom: 20,
+    textAlign: 'center',
+  },
+  optionsContainer: {
+    flexDirection: 'row',
+    justifyContent: 'space-around',
+  },
+  option: {
+    alignItems: 'center',
+    padding: 12,
+    borderRadius: 12,
+    minWidth: 80,
+  },
+  optionPressed: {
+    opacity: 0.7,
+  },
+  optionText: {
+    marginTop: 8,
+    fontSize: 14,
+  },
+});
