+++ styles/components/messageBubble.styles.ts
@@ -1,0 +1,128 @@
+import { StyleSheet } from 'react-native';
+
+export const messageBubbleStyles = StyleSheet.create({
+  container: {
+    marginVertical: 4,
+    maxWidth: '80%',
+  },
+  selfContainer: {
+    alignSelf: 'flex-end',
+  },
+  otherContainer: {
+    alignSelf: 'flex-start',
+  },
+  bubble: {
+    borderRadius: 16,
+    padding: 12,
+    elevation: 1,
+    shadowOffset: { width: 0, height: 1 },
+    shadowOpacity: 0.1,
+    shadowRadius: 1,
+  },
+  selfBubble: {
+    borderBottomRightRadius: 4,
+  },
+  otherBubble: {
+    borderBottomLeftRadius: 4,
+  },
+  messageText: {
+    fontSize: 16,
+  },
+  selfMessageText: {
+    color: '#000000',
+  },
+  footer: {
+    flexDirection: 'row',
+    justifyContent: 'flex-end',
+    alignItems: 'center',
+    marginTop: 2,
+  },
+  timeText: {
+    fontSize: 11,
+    opacity: 0.7,
+  },
+  media: {
+    width: 250,
+    height: 200,
+    borderRadius: 12,
+  },
+  videoContainer: {
+    width: 250,
+    height: 200,
+    borderRadius: 12,
+    overflow: 'hidden',
+    backgroundColor: '#F0F0F0',
+  },
+  videoThumbnail: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: 'rgba(0,0,0,0.1)',
+  },
+  videoInfo: {
+    position: 'absolute',
+    bottom: 0,
+    left: 0,
+    right: 0,
+    padding: 8,
+    backgroundColor: 'rgba(0,0,0,0.7)',
+  },
+  videoText: {
+    fontSize: 14,
+    color: '#FFFFFF',
+  },
+  videoSize: {
+    fontSize: 12,
+    color: '#FFFFFF',
+    opacity: 0.7,
+  },
+  audioContainer: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    padding: 8,
+    backgroundColor: '#000000',
+    borderRadius: 8,
+    marginTop: 4,
+    gap: 8,
+  },
+  audioText: {
+    flex: 1,
+    fontSize: 14,
+  },
+  fileContainer: {
+    width: 250,
+    borderRadius: 12,
+    overflow: 'hidden',
+    backgroundColor: '#F0F0F0',
+    flexDirection: 'row',
+    alignItems: 'center',
+    padding: 8,
+    marginBottom: 8,
+  },
+  fileText: {
+    flex: 1,
+    marginLeft: 8,
+    fontSize: 14,
+  },
+  modalContainer: {
+    flex: 1,
+    backgroundColor: '#000000',
+    justifyContent: 'center',
+  },
+  closeButton: {
+    position: 'absolute',
+    top: 40,
+    right: 20,
+    zIndex: 1,
+    padding: 8,
+  },
+  fullScreenVideo: {
+    width: '100%',
+    height: '100%',
+  },
+  footerActions: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: 4,
+  },
+});
