+++ styles/components/messageSearch.styles.ts
@@ -1,0 +1,68 @@
+import { StyleSheet } from 'react-native';
+
+export const messageSearchStyles = StyleSheet.create({
+  container: {
+    flex: 1,
+  },
+  searchBar: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    paddingHorizontal: 16,
+    paddingVertical: 8,
+    borderBottomWidth: StyleSheet.hairlineWidth,
+  },
+  input: {
+    flex: 1,
+    fontSize: 17,
+    marginHorizontal: 8,
+    paddingVertical: 8,
+  },
+  resultsList: {
+    paddingTop: 8,
+  },
+  resultItem: {
+    paddingHorizontal: 16,
+    paddingVertical: 12,
+    borderBottomWidth: StyleSheet.hairlineWidth,
+  },
+  resultContent: {
+    flex: 1,
+  },
+  messageInfo: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    marginBottom: 4,
+  },
+  chatName: {
+    fontSize: 16,
+    fontWeight: '600',
+    flex: 1,
+    marginRight: 8,
+  },
+  messageText: {
+    fontSize: 15,
+    lineHeight: 20,
+  },
+  timestamp: {
+    fontSize: 12,
+  },
+  loader: {
+    marginTop: 20,
+  },
+  error: {
+    color: '#FF3B30',
+    textAlign: 'center',
+    margin: 16,
+  },
+  noResultsContainer: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    paddingBottom: '20%',
+  },
+  noResults: {
+    marginTop: 12,
+    fontSize: 15,
+  },
+});
