+++ styles/components/searchModal.styles.ts
@@ -1,0 +1,33 @@
+import { StyleSheet } from 'react-native';
+
+export const searchModalStyles = StyleSheet.create({
+    container: {
+        flex: 1,
+    },
+    safeArea: {
+        flex: 1,
+    },
+    header: {
+        flexDirection: 'row',
+        alignItems: 'center',
+        paddingHorizontal: 8,
+        paddingVertical: 8,
+        borderBottomWidth: StyleSheet.hairlineWidth,
+    },
+    backButton: {
+        padding: 8,
+        marginRight: 8,
+    },
+    title: {
+        fontSize: 17,
+        fontWeight: '600',
+        flex: 1,
+        textAlign: 'center',
+    },
+    placeholder: {
+        width: 40,
+    },
+    content: {
+        flex: 1,
+    },
+});
