+++ styles/screens/loginScreen.styles.ts
@@ -1,0 +1,23 @@
+import { StyleSheet } from 'react-native';
+export const loginScreenStyles = StyleSheet.create({
+    safeArea: {
+      flex: 1,
+    },
+    container: {
+      flex: 1,
+      paddingTop: 20,
+    },
+    header: {
+      alignItems: 'center',
+      padding: 20,
+      marginBottom: 20,
+    },
+    subtitle: {
+      marginTop: 10,
+      fontSize: 16,
+      color: '#8F8F8F',
+    },
+    listContainer: {
+      paddingBottom: 20,
+    },
+});
